Customers with same cif

+++ Presentation/CMdm.UI.Web/Controllers/PhoneValidationController.cs
@@ -161,7 +161,7 @@ namespace CMdm.UI.Web.Controllers
             {
                 var ids = selectedIds
                     .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
-                    .Select(x => Convert.ToInt32(x))
+                    .Select(x => Convert.ToInt64(x))
                     .ToArray();
                 docs.AddRange(_dqQueService.GetPhoneValidationbyIds(ids));
             }
