+++ Services/CMdm.Services/CustomModule/Fcmb/IPhoneValidationService.cs
@@ -16,13 +16,13 @@ namespace CMdm.Services.CustomModule.Fcmb
         /// </summary>
         /// <param name="recordId">recordId identifier</param>
         /// <returns>Vendor</returns>
-        IList<PhoneValidation> GetPhoneValidationbyIds(int[] recordIds);
+        IList<PhoneValidation> GetPhoneValidationbyIds(Int64[] recordIds);
         /// <summary>
         /// Gets a Queitem by item reference identifier
         /// </summary>
         /// <param name="recordId">que identifier</param>
         /// <returns>Vendor</returns>
-        PhoneValidation GetPValbyId(int recordId);
+        PhoneValidation GetPValbyId(Int64 recordId);
         /// <summary>
         /// Gets all items
         /// </summary>
